fix: inject Function.prototype.bind polyfill into the app
There must be a more elegant way, but that was the only one that assured that the polyfill gets executed before any other script gets loaded. casper.options.clientScripts does not do the trick, the injected scripts get added after the existing ones. I think this is a common problem and there must be a better way to do it, but this will do for now