fix(env): manually add the test app to the paths that require looks up
This is necessary because the tested module gets `npm link`ed into the test
app. Inside this linked component `require` or `require.resolve` calls can
not know that they're symlinked and so they don't include peerDepdencies that
live in the app.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm